+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Added `ErrNilPointer` error
 
+### Fixes
+
+* Added nil pointer checks on all Env methods
+
 ## v2.0.0
 
 Released Monday, 7th October 2019.